# apps/api/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/organizations/{organizationId}/users/employees": {
      "get": {
        "description": "Retrieve employees users within an organization.",
        "operationId": "GetEmployeesUsersController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": { "default": 1, "type": "number" }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Number of items per page for pagination",
            "schema": { "default": 10, "type": "number" }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Order of the results",
            "schema": { "enum": ["asc", "desc"], "type": "string" }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Field to order the results by",
            "schema": { "enum": ["createdAt", "updatedAt", "email"], "type": "string" }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter results",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with user details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetEmployeesUsersResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get employees users",
        "tags": ["Users"]
      }
    },
...
  },
  "info": {
    "title": "Queue management",
    "description": "A queue management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "NotFoundErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "const": 404 },
          "message": { "type": "string" }
        },
        "required": ["statusCode", "message"]
      },
...
      "UnauthorizedErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "const": 401 },
          "message": { "type": "string" }
        },
        "required": ["statusCode", "message"]
      },
      "GetEmployeesUsersResponseDto_Output": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "name": { "type": "string" },
                "email": { "type": "string" },
                "emailVerified": { "type": "boolean" },
                "banned": { "type": "boolean" },
                "banReason": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                "banExpires": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                "image": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                "role": { "type": "string" },
                "createdAt": { "type": "string" },
                "updatedAt": { "type": "string" }
              },
              "required": [
                "id",
                "name",
                "email",
                "emailVerified",
                "banned",
                "banReason",
                "banExpires",
                "image",
                "createdAt",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          },
          "total": { "type": "number" }
        },
        "required": ["users", "total"],
        "additionalProperties": false
      },
...
    }
  }
}